Update EmitterProperties.cs class fields
Add private const float fields.
Remove private _Direction and _Position fields.
Rename SetAndGet() methods to Get().
Add local float variable to GetDirection() and GetPosition() methods.
Refactor GetDirection() methods to use const fields.
Update CubeEmitter.cs references to EmitterProperties.cs methods.
Update CubeHandler.cs SetDirection() method to use float arguments.

+++ Shapes/Assets/Scripts/CubeHandler.cs
@@ -30,10 +30,10 @@ public class CubeHandler : MonoBehaviour
         _boundaryWrapDistance = GetComponent<Collider>().bounds.size.x * 1.1f;
     }
 
-    public void SetDirection(int x, int y)
+    public void SetDirection(float x, float y)
     {
-        _horizontal = (x > 0) ? Random.Range(0.0f, 1.0f) : Random.Range(-1.0f, 0.0f);
-        _vertical = (y > 0) ? Random.Range(0.0f, 1.0f) : Random.Range(-1.0f, 0.0f);
+        _horizontal = (x > 0.0f) ? Random.Range(0.0f, 1.0f) : Random.Range(-1.0f, 0.0f);
+        _vertical = (y > 0.0f) ? Random.Range(0.0f, 1.0f) : Random.Range(-1.0f, 0.0f);
     }
 
     private void RecalculateDirection()
